semi-enforce notebook tests to pass for non-spotdev user
Spot deployment script requires a command line argument to start the deployment
process. This parameter is a proxy for enforcing that notebooks be tested by a
non spotdev user before deployment. The parameter indicates that the template
notebooks have passed successfully with a user that is not in the spotdev
group. The dev notebooks should be tested with two types of users before
deployment, but automating that from a single user shell is not
straight-forward.